# Improve performance of atomic load in toplevel code