As noted in https://github.com/JuliaLang/julia/issues/47561, the performance of toplevel code is pretty bad because of the atomic barrier (loading the global world age into the task-local one) that's emitted after every instruction now. @vtjnash noted that monotonic ordering is sufficient, which recovers some performance (1.5s -> 0.7s)

#45550 also moved the atomic load from `emit_stmtpos`, where it doesn't trigger after every statement type, to one level up where it does emit a load unconditionally. @vtjnash was that intended, or did you only move it because of the removal of `ctx.world_age_field`? Moving it back speeds up https://github.com/JuliaLang/julia/issues/47561 another bit (0.7s -> 0.2s).

Of course, the real fix for https://github.com/JuliaLang/julia/pull/47561 is https://github.com/JuliaLang/julia/pull/47576, but it can't hurt to improve performance of toplevel code a bit.